Add files via upload
the file maze1.txt for representing a 5-by-5 maze with no wall, and the turtle in the middle. 

the file maze2.txt for representing a 5-by-5 maze with walls on all the borders and an outlet in the South. The turtle starts in the middle.

the file maze3.txt for representing a 5-by-5 maze which leads to maximize the number of recursive calls to search_from.